Update welcome.yml
Update to use GitHub CLI (gh) to add a comment when a pull request is opened.  To allow GitHub CLI to post a comment, we set the GITHUB_TOKEN environment variable to the value of the GITHUB_TOKEN secret, which is an installation access token, created when the workflow runs. For more information, see "Automatic token authentication." We set the PR_URL environment variable to the URL of the newly created pull request, and we use this in the gh command.